refactor: 챌린지 관련 인터페이스를 ai-challenge.types.ts로 이동
